Update the explain for Dynamic Bitmap Index Scans
Currently for plans generated by ORCA, the explain plan prints out the
name of the first partitioned index, regardless of whether or not it is
the correct index. This may be misleading because the executor actually
looks at the entire index and properly selects only the indexes needed for
the required partitions.

The first partitioned index's OID is used to help the executor find the
indexes needed. In order to prevent future misgivings, do not print the
index name here in the explain plans for partition tables in ORCA.

This does not affect explains by planner or other plans as they do not
follow the same code path.

Co-authored-by: Ashuka Xue <axue@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>