6X Backport: Fix dispatch of SET TIME ZONE INTERVAL statement.
Backported from: fe3f58a9dbc1da0fb5153a19e255b7969295f8bb

This fixes Issues: [1]#9055 and incidentally fixes related issue [2]#9044.

Tl;dr: Instead of dispatching `SET TIME ZONE INTERVAL <args>` when this
flavor was used, we extracted the arguments following `INTERVAL` and
dispatched `SET timezone TO <args>`.

Please refer to issue #9055 for the repro, complete RCA and fix details.

Additional Notes:

1. While working on these two issues, I saw that `DispatchSetPGVariable()`
became out of date with respect to upstream code. This is because it has
a foreach loop to add the argument list to the dispatch string that is
borrowed heavily from `flatten_set_variable_args()`. In fact that is the
precise reason for related issue: #9044. We did not port the changes
brought by upstream commit: 6fff5c3. So I ported those changes over.

2. I added a test for this scenario and I refactored some existing tests
in this area for them to be more explicit in their intent.

[1] https://github.com/greenplum-db/gpdb/issues/9055
[2] https://github.com/greenplum-db/gpdb/issues/9044