packcore: list libraries with ld-linux.so when ldd is unavailable
When the gdb command is not found on the system, packcore could use the
ldd command to list the shared libraries.  When ldd is neither found now
we will try to list the libraries with the `ld-linux.so` directly.

(cherry picked from commit ab56489f715cd6258c51aaae1488101560fd1ddb)