Add check and test for upgrading table with unique/primary constraint (#9190)
For unique or primary key constraint, the index name is auto generated, and if
the default index name is already taken by other objects, an incremental number
is appended to the index name.  This means that pg_upgrade cannot upgrade a
cluster containing indexes of such type, they must be handled manually
before/after the upgrade. Although, the issue exists only with such indexes,
but we wholesale ban upgrading of unique or primary key constraints. Such,
constraints must be dropped before the upgrade, and can be recreated after the
upgrade.

Co-authored-by: Bhuvnesh Chaudhary <bhuvnesh2703@gmail.com>
Co-Authored-By: David Kimura <dkimura@pivotal.io>