When serializing a tuple for Motion, don't decompress compressed datums.
If a datum is toasted and compressed, only detoast it before serializing.
No need to decompress, the receiver can decompress it if needed. That's
a big win, if the receiver is just going to store the value back to disk,
and doesn't need to decompress it at all.

The corresponding codepath for MemTuples was already doing this, within
memtuple_form_to().

Addresses github issue https://github.com/greenplum-db/gpdb/issues/9253

Reviewed-by: Asim R P <apraveen@pivotal.io>