COPY: Simplify partition distribution policy lookup
Commit a8aa1c4acc95ea42 introduced a bug when a partition has a
different distribution policy (in terms of column numbers) from the base
table: we would perform a partition lookup -- using a tuple table slot
for the partition but (incorrectly) with a tuple descriptor for the base
table. This would sometimes lead to an error of "no partition for
partitioning key". Upon closer inspection, we didn't even need to look
up the partition because the caller already knew!

This commit fixes that, and simplifies the logic in
GetDistributionPolicyForPartition to just take a resultRelInfo instead
of the tuple values.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>
Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>
Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>