Set the default value of wal_keep_segments to 5 (#9276)
In this PR(https://github.com/greenplum-db/gpdb/pull/9248),
we set the default value of wal_keep_segments to 0, the same
as upstream, because we have replication slot to avoid removal
of WAL files required by the mirror. It seems to be fine.

But there is no replication slot for master/standby now. It's
unsafe to remove WAL files if the file was required by the standby.
So, for now, before the replication slot is added to master, let's
set the default value of wal_keep_segments to 5.

(cherry picked from commit 3ce78553e3b5de66c4a196eb727bcc4899c0ea8b)