Change default value of wal_sender_timeout GUC
Based on reports from field for GPDB, 1 min of wal_sender_timeout GUC
is causing primary to terminate the replication connection too often
in heavy workload situations. This causes mirror to be marked down and
piles up WAL on primary. This is moslty seen in configurations where
fsync takes long time on mirrors. Hence, would be helpful to have
higher default value of this GUC to avoid unnecessary marking mirror
down situations. Only downside of this change would be when connection
between primary and mirror exist but for some reason mirror doesn't
respond, it will be detected little later compared to previous 1 min
timeout. But 1 min timeout is causing major downside and mirrors need
to be manually recovered after being marked down. Hence, its desirable
to not falsely break the connectiion due to timeout.

Increasing the timeout to 5 mins is just a educated guess as its hard
to come up with reasonable default, but bumping the value is desired
based on inputs.

Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>