Fix crash with JSON/YAML format EXPLAIN.
The code in EXPLAIN that displays the gang information of a node was
correctly prepared to handle the case that a node was missing flow
information, by looking at the child plan's flow instead. However, it's
possible to have two such nodes on top of each other. We need to look at
the grandchild's flow in that case.

This only occurs with JSON/YAML format EXPLAIN, because in text format we
only print the gang information on Motion nodes.

Fixes https://github.com/greenplum-db/gpdb/issues/9359. Fix on 6X_STABLE
only; 5X_STABLE didn't have JSON/YAML format output, and this works on
master. I'm not entirely sure why this works on master, but I'm not going
to spend time figuring that out right now, because I'm just about to
refactor this so that it's not based on the Flow nodes at all
(https://github.com/greenplum-db/gpdb/pull/9093).

Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>