Fix another crash in JSON/YAML-format EXPLAIN, due to missing Flow.
This is similar to commit 51257848, but this time it was a Bitmap Index
Scan node that was missing the Flow information. Fix by populating the
Flow for Bitmap Index Scans.

This class of bugs seem to happen pretty often, so make the explain code
more defensives so that it does not crash if this happens again. In both
of these bugs, the Flow information is only needed so that we can print
the # of segments in the EXPLAIN output. That's not critical, so in
production, it seems better to print possible incorrect information than
crash. Keep the Assert, though, so that we still catch these in
development.

Fixes github issue https://github.com/greenplum-db/gpdb/issues/9404.
Commit to 6X_STABLE, like the previous issue.

Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>
Reviewed-by: Ning Yu <nyu@pivotal.io>