Do not dispatch DISCARD ALL and emmit a NOTICE if it is used in the dispatcher
DISCARD ALL can not run inside a transaction and as a consequence, it is not
possible to use two-phase commit while dispatching. As a consequence, it is not
possible to recover from segment errors which can potentially leave the cluster
in an unusable state.

Do not attempt to dispatch DISCARD ALL. Rather, emmit a NOTICE when it is used
from the dispatcher. It is assumed that if the commands reaches the segments,
then human/utility mode intervesion has occured.

NEEDS documentation changes.

Backport of: https://github.com/greenplum-db/gpdb/pull/9387

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>