Make READ_INT/OID_ARRAY() macros in readfuncs.c and readfast.c compatible.
One of them took 'count' as an argument, and the other took 'field name'.
Harmonize on plain 'count'.

This is different from the definitions of these macros in later versions
of PostgreSQL, so we should revisit this later. But for now, in 6X_STABLE,
let's standardize on the the readfast.c version, because there are more
callers of these macros in readfast.c.