Fix node out/read routines for Partition-related structs.
The functions for PartitionRule in outfuncs.c and readfuncs.c were missing
the 'parisdefault' field, and readfuncs.c function for PartitionNode was
missing the 'default_part' field. Those omissions were harmless, because
those functions are in fact unused; we never materialize PartitionRule
or PartitionNode structs to disk. But let's be tidy.

To fix, add the missing fields to the out/readfuncs.c functions, and to
reduce the chance of similar bugs in the future and to reduce the amount
of duplication, remove binary versions of these functions from
outfast.c/readfast.c.

Also remove the binary versions of _readPartition() and _outPartition()
while we're at it. No fields were missing from those, but less duplicated
code is good.

The string and binary versions for Partition and PartitionRule were not
exactly the same. The binary version wrote out the 'parruleord' and
'parnatts' fields as 16-bit integers, whereas the string version used a
32-bit integer. But that's OK, we're free to change the binary functions
in minor versions, and had in fact already made one such change since
the last minor version (6.3.0), in commit 22b72301. Using a 16-bit int
is in principle more compact, but these are not that space-critical that
it would matter.

Discussion: https://github.com/greenplum-db/gpdb/issues/9390
Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>