Fix confusion with dropped columns in ANALYZE dispatching.
The internal gp_acquire_sample_rows() function was confused on whether
the bitmap for oversized columns includes dropped columns or not. It does
not, but the loop that populated it looked at the 'attisdropped' flag
anyway. That was bogus, because in this loop, the dropped columns have
already been excluded from the 'natts' column count.

Also, in the QD side, the code that used the oversized-column bitmap was
confused on the indexing of the colLargeRowIndexes array. It's indexed
by attribute number - 1, but the 'i' variable in the loop was referring
to the index in the vacattrstats[] array. They are not the same, if some
columns are not analyzed, e.g. because some columns or dropped or are of
an un-analyzable type.

Fixes github issue https://github.com/greenplum-db/gpdb/issues/9503.

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>