Allow cluster to start when standby host is down
Previously, gpstart could not start the cluster if a standby master host
was configured but currently down.  In order to check whether the
standby was supposed to be the acting master (and prevent the master
from being started if that was the case), gpstart needed to access the
standby host to retrieve the TimeLineID of the standby, and if the
standby host was down the master would not start.

This commit modifies gpstart to assume that the master host is the
acting master if the standby is unreachable, so that it never gets into
a state where neither the master nor the standby can be started.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>
Co-authored-by: Mark Sliva <msliva@pivotal.io>
Co-authored-by: Adam Berlin <aberlin@pivotal.io>
(cherry picked from commit 29c759ab8c1f4179e46b51c91a808e76f6747075)