Modify CI infrastructure and remove no longer needed pipelines/scripts/files
The main change here is that we completely get rid of the ORCA prod
pipeline, as there's no need to create a tag/publish artifacts anymore.

We also now build and test the orca unit tests during the compile step.

Additionally, I've refactored the orca-dev "slack command" pipeline to
use the canonical compile_gpdb.bash script and based the pipeline off
the existing PR pipeline. This simplifies the `test_orca_pipeline.yml`
file significantly and it now doesn't rely on custom scripts.

There are still quite a few files in the concourse directory that are
used by the `test_explain_pipeline.yml` and a couple of pipelines that
are external to this repo. Some of these aren't actively being used and
need to be cleaned up/removed, but I'll leave that for a later time.

Also, uses https for xerces-c download and checks hash to ensure correct
download.