Fix configure and cmake to build ORCA with debug
Previously, we used a config file that was modified by cmake. Instead,
use configure to modify macros. Additionally use a compile definition in
cmake so we don't need a separate config file in gpos.

This also enables compile-time warnings in the src/backend/gporca files. C++ files
in the translator never had these warnings enabled, and so are not
enabled in src/backend/gpopt. A block of unused code is removed to get
pass the compile warnings.