Fix mis-merge of skew hashtable reset before advancing batch
Resetting the skew hashtable-related variables should always happen
after batch 0. This was a mis-merge through which the skew hashtable
reset code was added as the "else" condition in an unrelated conditional.

In rare cases the "if" condition would be hit instead, causing a segfault
upon retrieving the next outer tuple when checking to see if the outer
tuple is in the skew hashtable in ExecHashGetSkewBucket().

This would only happen, however, for plans in which a fast path rescan
hint has been explicitly set in a sub-tree containing a hashjoin, and,
specifically, a hashjoin which spills. This condition will only trigger
when the skew hashtable is in use, which will only happen on segments
with local statistics for the table in question.

(cherry picked from commit d3ccff0434e4d005aefcaa8c6b9f7b0f770718c2)