Refactor and correct prefetch joinqual
We introduces prefetch joinqual logic in commit fa762b and
later refactoed it in commit 36a93ba. But the logic before does
not handle things correct. Previously we finally check this until
ExecInitXXX and forgot to check if there is joinqual.

This commit fixes the issue for planner. It determines the bool
field `prefetch_joinqual` during planning stage.  The logic is
first set the field to show the risk in create_xxxjoin_plan, and then
in create_join_plan after we get the plan, if there is risk then we
loop each joinqual to see if there is motion in it (The only case
is the joinqual is a SubPlan). In 6X we add motions for SubPlan
after we have gotten the plan. So at the time of this function calling,
we do not know if the SubPlan contains motion. We might walk the plannedstmt
to accurately determine this, but that makes the code not that clean.
So, we adopt an over kill method here: any SubPlan contains motion. At least,
this can avoid motion deadlock.

Also this commit improves the logic of ExecPrefetchJoinQual to
solve the github issue:  https://github.com/greenplum-db/gpdb/issues/8677
It fixes this by flattening the bool expr of the joinqual and force each
of them to be Prefetched.