Fix a couple of ORCA assertions
These were exposed when running ICW with ORCA asserts enabled.

In DeriveJoinStats, EopLogicalFullOuterJoin is also a valid logical join
operator. In IDatum, we need to check that doubles are within some
epsilon as we're not passing in the full 64 bit IEEE value to ORCA.

With fixing the assertion, we would need to regenerate the mdp for
MinCardinalityNaryJoin However, there is no DDL/query for this test so
it is difficult to update. Since it also didn't seem to provide much
value, we're removing it.