Fix the way we check if a partition table is empty in Relcache translator
The relation statistics object stores a flag which indicates if a relation is
empty or not. However this was being set based on the value of stats_empty
flag. This flag was being set in the function `cdb_estimate_partitioned_numtuples`.
It would be set to true if any of the child partitions had their relpages set
to 0. But if other partitions had data in them, then the relation should not be
considered empty. We should actually be checking for the total number of rows
to determine if a partition table is empty or not. Also the stats_empty flag
was not being used for anything else so this commit removes it.