Return early if toast table is not required
While attempting to create a toast table in case of binary upgrade,
preassigned oids are required to see if a toast table is required.
Currently, in the code, the check to see if there is AccessExclusiveLock
is executed early even though a toast table is not required. In case,
there is a no preassigned oid to create a toast table, we should bail
out early as no AccessExclusiveLock would have been taken on the
relation.