Convert disable_cost to a guc
There was a discussion about disable_cost being small sometimes on upstream,
https://www.postgresql.org/message-id/flat/CAO0i4_SSPV9TVxbbTRVLOnCyewopcc147fBZy%3Df2ABk15eHS%2Bg%40mail.gmail.com
but there is no conclusion there although many solutions are discussed there.
This issue seem to be more urgent for Greenplum since Greenplum is a
distributed system and can handle much more data than single node postgres.

Recently we encountered a user issue again that is quite relevant of this. In
the case merge join, instead of hashjoin is chosen. Admittedly in the case
there is a corner costing issue (might be uglily fixed though) in the hashjoin
code, but generally we are wondering if we should tune disable_cost to fix
given small disable_cost value has been known to be an issue sometimes also.

We expect finally disable_cost should be gone on gpdb master, and we know we
need to work to make hashjoin costing more accurate (though sometimes it is
hard), but for now users are waiting for a solution on gpdb6, so we use the guc
solution temporarily. Note in real scenario please temporarily tune it only
when needed (typically when enabled path cost is near to or greater than
disable cost but we want the enabled path) else there might be some side effect
of planning for "disabled" paths due to precision (e.g. some paths are ignored
due to precision).

Co-authored-by: Jinbao Chen <jinchen@pivotal.io>

Reviewed-by: Ning Yu <nyu@pivotal.io>

Cherry-picked from 9110839846b8ffa273cbf03cd79515ab6c4fe67f