Handle disabling of bitmap and bpchar_pattern_ops indexes in pg_upgrade
In GPDB6, the access method oid is 7013 whereas it's 3013 in GPDB5.
Until now, the bitmap indexes were not marked invalid as the query to
update the catalog relied on incorrect OID.  This commit removes the
hardcoded oid, and instead uses a catalog query.

Also, GPDB uses the GUC allow_system_table_mods to allow updates on
catalog tables. While disabling the indexes with bpchar_pattern_ops
indexes, the GUC was not set and the query used to fail. This commit
fixes it too.

Since upgraded target master database is copied to segments, reset the
index status to valid where required. The queries to ensure that all the
objects have been migrated from old segment to new segments excludes
invalid objects and will complain that index has not been migrated. The
index will be marked invalid during upgrade of the segment.