Increase instance memory for gpexpand tests
The gpepxand tests were going OOM with ORCA in debug build, as the
segments require more memory. Previously, ORCA was not being run in
debug build so this was not an issue. We've bumped up the instance from
n1-standard-2 to n1-standard-4, which doubles our memory from 7.5GB to
15GB.