Fix memory leak and bug in checkpointer process (#9766)
1. The CurrentMemoryContext in CreateCheckPoint is long-live
memory context owned by the checkpointer process. The memory context
is reset only if the error occurs. So memory leak will lead to
huge memory leak in the OS.
2. The prepared transactions are not appended as an extension in
the checkpoint wal record, which introduces a bug. It occurs:
  1) T1 does some DML and is prepared.
  2) T2 runs a checkpoint.
  3) The seg0 crashes before the transaction T1 successfully runs `COMMIT PREPARE`, but
     all other segments have successfully committed.
  4) When running local recovery, seg0 doesn't know T1 is prepared
     and needs to be committed again.
This is different from the master branch. Master uses files to record
whether there exists any prepared transaction and what them are.

Reviewed-by: Ning Yu <nyu@pivotal.io>
Reviewed-by: Gang Xiong <gxiong@pivotal.io>