Fix another missing-flow bug in JSON-format EXPLAIN.
Similar to 3bdece19da and 512578487b, an Append node doesn't always have
a Flow attached to it. Teach EXPLAIN code to dig into its children.

This seems to be a recurring issue, but since it's been refactored away in
'master', I guess we'll just whack these on 6X_STABLE as they're reported.
Downgrade it to a WARNING, though; an assertion failure seems too harsh
for this. Now you get the warning even on non-assertion builds, but I
think that's fine.

Fixes https://github.com/greenplum-db/gpdb/issues/9819

Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>